+++ src/Twilio/Rest/Api/V2010/Account/Conference/ParticipantList.php
@@ -88,6 +88,13 @@ class ParticipantList extends ListResource {
             'CallReason' => $options['callReason'],
             'RecordingTrack' => $options['recordingTrack'],
             'TimeLimit' => $options['timeLimit'],
+            'MachineDetection' => $options['machineDetection'],
+            'MachineDetectionTimeout' => $options['machineDetectionTimeout'],
+            'MachineDetectionSpeechThreshold' => $options['machineDetectionSpeechThreshold'],
+            'MachineDetectionSpeechEndThreshold' => $options['machineDetectionSpeechEndThreshold'],
+            'MachineDetectionSilenceTimeout' => $options['machineDetectionSilenceTimeout'],
+            'AmdStatusCallback' => $options['amdStatusCallback'],
+            'AmdStatusCallbackMethod' => $options['amdStatusCallbackMethod'],
         ]);
 
         $payload = $this->version->create('POST', $this->uri, [], $data);